+++ examples/components/Prerendered.js
@@ -9,11 +9,15 @@ class Prerendered extends Component {
         kiosk: PropTypes.object.isRequired
     };
 
+    static defaultProps = {
+        initialOpen: false
+    };
+
     constructor(props) {
         super(props);
 
         this.state = {
-            isOpen: props.initialOpen || false
+            isOpen: props.initialOpen
         };
     }
 
@@ -30,12 +34,14 @@ class Prerendered extends Component {
     render() {
         return (
             <div>
-                <a href="" className="btn btn-primary" onClick={this.toggleModal.bind(this)}>Open</a>
+                <a className="btn btn-primary" onClick={::this.toggleModal}>Open</a>
 
-                <Modal isOpen={this.state.isOpen}
-                       title="Longcat is long"
-                       kiosk={this.props.kiosk}
-                       onRequestClose={this.toggleModal.bind(this)}>
+                <Modal
+                    isOpen={this.state.isOpen}
+                    title="Longcat is long"
+                    kiosk={this.props.kiosk}
+                    onCancel={::this.toggleModal}
+                >
                     <p>
                         Viral deep v squid chia, letterpress wayfarers artisan
                         meggings tote bag four loko keffiyeh hoodie cronut four
@@ -51,7 +57,7 @@ class Prerendered extends Component {
 
                     <p>
                         Helvetica banh mi Wes Anderson aesthetic, stumptown keytar
-                        ugh beard VHS you probably haven't heard of them Carles
+                        ugh beard VHS you probably haven not heard of them Carles
                         Neutra bespoke pour-over Odd Future. Crucifix banjo Tumblr
                         3 wolf moon, readymade tilde master cleanse. Neutra fingerstache
                         seitan, cornhole hoodie Pinterest messenger bag food
@@ -63,7 +69,7 @@ class Prerendered extends Component {
                         Fashion axe paleo selfies, put a bird on it small batch artisan
                         salvia lomo messenger bag polaroid synth leggings drinking
                         vinegar letterpress organic. Slow-carb church-key chambray
-                        disrupt, Portland you probably haven't heard of them Schlitz
+                        disrupt, Portland you probably haven not heard of them Schlitz
                         literally PBR&amp;B trust fund fap.
                     </p>
 
@@ -81,26 +87,26 @@ class Prerendered extends Component {
                         Literally Williamsburg butcher, small batch drinking vinegar
                         bicycle rights messenger bag. Readymade 3 wolf moon blog ennui
                         mumblecore selvage sartorial plaid, tousled fap paleo. You probably
-                        haven't heard of them YOLO actually, leggings four dollar toast
+                        haven not heard of them YOLO actually, leggings four dollar toast
                         street art kale chips Kickstarter 8-bit messenger bag. Artisan
                         Truffaut actually beard, authentic vegan pour-over tattooed Marfa
                         tote bag narwhal try-hard. Blog gluten-free single-origin coffee
-                        crucifix pug, Pitchfork you probably haven't heard of them distillery
+                        crucifix pug, Pitchfork you probably haven not heard of them distillery
                         master cleanse meggings cardigan. Brooklyn artisan whatever food truck,
                         Carles tilde pug tofu pour-over put a bird on it Banksy migas you
-                        probably haven't heard of them mlkshk flannel. Hoodie mlkshk four
+                        probably haven not heard of them mlkshk flannel. Hoodie mlkshk four
                         dollar toast sriracha, Brooklyn Bushwick pug 8-bit.
                     </p>
 
                     <p>
-                        Blog typewriter you probably haven't heard of them locavore,
+                        Blog typewriter you probably haven not heard of them locavore,
                         letterpress twee authentic. Yr distillery post-ironic, ennui irony
                         American Apparel literally Tumblr. Kickstarter +1 chillwave sartorial
                         distillery, normcore Carles single-origin coffee American Apparel flannel.
                         Authentic pour-over stumptown forage, cray direct trade literally Pinterest
                         locavore 3 wolf moon organic slow-carb vegan. Mlkshk dreamcatcher try-hard,
                         butcher Blue Bottle cred PBR typewriter bespoke. Listicle ennui pork belly
-                        sriracha, Bushwick meggings letterpress DIY butcher McSweeney's slow-carb
+                        sriracha, Bushwick meggings letterpress DIY butcher McSweeney is slow-carb
                         forage direct trade. Carles jean shorts VHS, chambray hashtag PBR Echo Park
                         authentic retro listicle ugh raw denim skateboard literally.
                     </p>
@@ -109,7 +115,7 @@ class Prerendered extends Component {
                         Banksy hoodie pop-up, Pitchfork skateboard DIY typewriter selvage
                         dreamcatcher. Art party hoodie XOXO, typewriter slow-carb ugh Odd
                         Future lo-fi mumblecore PBR&amp;B letterpress stumptown Brooklyn normcore
-                        viral. You probably haven't heard of them Neutra plaid Shoreditch ethical,
+                        viral. You probably haven not heard of them Neutra plaid Shoreditch ethical,
                         kitsch fashion axe. IPhone deep v Intelligentsia, +1 squid Pinterest
                         Williamsburg gentrify selvage Bushwick chambray master cleanse mixtape
                         Godard polaroid. Stumptown listicle butcher Echo Park PBR&amp;B Brooklyn