+++ server.js
@@ -1,3 +1,1 @@
-require('babel/register');
-
 require('./examples/server');