+++ src/index.js
@@ -1,3 +1,10 @@
 import Modal from './browser';
 
+import ModalBody from './components/dom/ModalBody';
+import ModalHeader from './components/dom/ModalHeader';
+
+
+Modal.Body = ModalBody;
+Modal.Header = ModalHeader;
+
 export default Modal;